add s21_floor() with tests

+++ include/s21_math.h
@@ -23,7 +23,7 @@ double s21_ceil(double x);
 long double s21_cos(double x);
 long double s21_exp(double x);
 double s21_fabs(double x);
-long double s21_floor(double x);
+double s21_floor(double x);
 long double s21_fmod(double x, double y);
 long double s21_log(double x);
 long double s21_pow(double base, double exp);
